backend/requirements.txt
fastapi
uvicorn
joblib
numpy
pandas
tensorflow
shap
scikit-learn
xgboost
httpx
python-dotenv
